# src/main/resources/swagger/project1.yml
openapi: 3.0.3
info:
  title: Servicio de Gestión Bancaria
  description: El proyecto gestiona información de clientes, productos financieros, y transacciones, abarcando cuentas bancarias, créditos y tarjetas.
  termsOfService: http://swagger.io/terms/
  contact:
    email: anacarenasr@sgmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo
...
paths:
...
  /account:
    get:
      tags:
        - account
      summary: Obtener lista de cuentas (Account)
      description: Retorna una lista de todas las cuentas registradas.
      responses:
        '200':
          description: Lista de cuentas obtenida exitosamente
          content:
            application/json:
              example:
                - id: 1
                  customerId: 1
                  typeAccount: savings
                  descriptTypeAccount: Cuenta de Ahorro
                  amount: 5000.00
                  maintance: 0.00
                  transaction: []
                  operationDay: null
                  dateAccount: '2022-01-20'
                  numberAccount: '123-456-789'
                  typeCustomer: individual
                - id: 2
                  customerId: 2
                  typeAccount: current
                  descriptTypeAccount: Cuenta Corriente
                  amount: 10000.00
                  maintance: 20.00
                  transaction: []
                  operationDay: null
                  dateAccount: '2022-01-20'
                  numberAccount: '987-654-321'
                  typeCustomer: business

    post:
      tags:
        - account
      summary: Crear una nueva cuenta (Account)
      description: Permite registrar una nueva cuenta en el sistema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        '201':
          description: Cuenta creada exitosamente
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: Error de validación en la solicitud
  /account/{id}:
    get:
      tags:
        - account
      summary: Obtener detalles de una cuenta (Account) por ID
      description: Retorna los detalles de una cuenta específica según su ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalles de la cuenta obtenidos exitosamente
          content:
            application/json:
              example:
                id: 1
                customerId: 1
                typeAccount: savings
                descriptTypeAccount: Cuenta de Ahorro
                amount: 5000.00
                maintance: 0.00
                transaction: []
                operationDay: null
                dateAccount: '2022-01-20'
                numberAccount: '123-456-789'
                typeCustomer: individual
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              example:
                message: Cuenta no encontrada

    put:
      tags:
        - account
      summary: Actualizar una cuenta (Account) existente
      description: Permite actualizar los detalles de una cuenta existente.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        '200':
          description: Cuenta actualizada exitosamente
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: Error de validación en la solicitud
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              example:
                message: Cuenta no encontrada

    delete:
      tags:
        - account
      summary: Eliminar una cuenta (Account)
      description: Permite eliminar una cuenta según su ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Cuenta eliminada exitosamente
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              example:
                message: Cuenta no encontrada
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
          description: ID del cliente asociado a la cuenta.
        typeAccount:
          type: string
          description: Tipo de cuenta.
        descriptTypeAccount:
          type: string
          description: Descripción del tipo de cuenta.
        amount:
          type: number
          format: float
          description: Monto disponible en la cuenta.
        maintance:
          type: number
          format: float
          description: Monto de mantenimiento de la cuenta.
        transaction:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
          description: Lista de transacciones asociadas a la cuenta.
        operationDay:
          type: string
          format: date
          description: Día de operación de la cuenta.
        dateAccount:
          type: string
          format: date
          description: Fecha de creación de la cuenta.
        numberAccount:
          type: string
          description: Número de cuenta.
        typeCustomer:
          type: string
          enum: [individual, business]
          description: Tipo de cliente asociado a la cuenta.

      required:
        - customerId
        - typeAccount
        - descriptTypeAccount
        - amount
        - maintance
        - transaction
        - operationDay
        - dateAccount
        - numberAccount
        - typeCustomer
...
    Transaction:
      type: object
      properties:
        productType:
          type: string
          description: Tipo de producto asociado a la transacción.
        productId:
          type: integer
          format: int64
          description: ID del producto asociado a la transacción.
        customerId:
          type: integer
          format: int64
          description: ID del cliente asociado a la transacción.
        transactionType:
          type: string
          description: Tipo de transacción.
        amount:
          type: number
          format: float
          description: Monto de la transacción.
        transactionDate:
          type: string
          format: date
          description: Fecha de la transacción.
        customerType:
          type: string
          enum: [individual, business]
          description: Tipo de cliente asociado a la transacción.
        balance:
          type: number
          format: float
          description: Saldo después de la transacción.

      required:
        - productType
        - productId
        - customerId
        - transactionType
        - amount
        - transactionDate
        - customerType
        - balance